+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   `--env-file` flag, files can be specified as a comma-separated list or
   by using the flag multiple times. Variables defined in later files take
   precedence.
+- `singularity.conf` now accepts setting the `allow uts ns` option, and can
+invalidate the use of the `--uts` and `--hostname` flags.
 
 ## 4.1.3 \[2024-05-08\]