+++ CHANGELOG.md
@@ -9,6 +9,12 @@
   by using the flag multiple times. Variables defined in later files take
   precedence.
 
+## Bug Fixes
+
+- `%files from` in a definition file will now correctly copy symlinks that point
+  to a target above the destination directory, but inside the destination stage
+  rootfs.
+
 ## 4.1.3 \[2024-05-08\]
 
 ### Requirements