+++ CHANGELOG.md
@@ -5,14 +5,14 @@
 ### Bug Fixes
 
 - Use ABI 3 for Apparmor profile on Ubuntu <23.10.
-
-### Bug Fixes
-
 - Avoid unnecessary copying / extraction of OCI images and Docker tarballs into
   a layout directory when they are directly accessible as a local file /
   directory.
 - Avoid unnecessary intermediate temporary image layout when building from
   Dockerfile to OCI-SIF.
+- `%files from` in a definition file will now correctly copy symlinks that point
+  to a target above the destination directory, but inside the destination stage
+  rootfs.
 
 ## 4.1.3 \[2024-05-08\]